+++ src/main/kotlin/balotenketo/balotaro/model/Poll.kt
@@ -1,23 +1,28 @@
 package balotenketo.balotaro.model
 
 import com.fasterxml.jackson.annotation.JsonIgnore
+import io.swagger.annotations.ApiModel
+import io.swagger.annotations.ApiModelProperty
 import org.joda.time.DateTime
 import org.springframework.data.annotation.Id
 import java.util.*
 
+@ApiModel
 data class Poll(
 
         @JsonIgnore
         val admin: String = "",
 
-        val secret: String = TokenGenerator.generate(),
-
+        @ApiModelProperty
         val createdDate: Date = DateTime.now().toDate(),
+
         val timeToLive: Long = 60 * 24 * 30, // In minutes
 
-        val choices: Map<String, String> = emptyMap()
+        @ApiModelProperty("Possible choices")
+        val choices: Set<String> = emptySet()
 ) {
 
     @Id
+    @ApiModelProperty("Id of the poll")
     lateinit var id: String
 }