+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * `/poll/create?choices={choices}` Create a new poll (authenticated users only)
 * `/poll/createTokens/{pollID}?count={count}` Create new tokens for a poll (Authenticated admin of the poll only)
 * `/poll/vote/{tokenID}/{secret}?ballot={ballot}` Submit a ballot and consume the vote-token
+* Swagger to document the API (UI at `/swagger-ui.html`)